add cookies.js
