+++ src/pages/_app.js
@@ -11,6 +11,7 @@ const theme = createTheme({
   }
 })
 function MyApp({ Component, pageProps }) {
+  
   return (
     <Provider store={store}>
       <ThemeProvider theme={theme}>
